chore: move ics files to root
