+++ ics-assessment/README.md
@@ -1,6 +1,7 @@
 # ICS Assessment Scripts
 
 Lightweight Python utilities for assessing ICS eligibility across three categories:
+
 - Engagement (governance participation and community activity)
 - Experience (solo-staking and validator/operator credentials)
 - Humanity (sybil-resistance and social presence)
@@ -11,22 +12,25 @@ The methodology, scoring and the sources are described in the corresponding [Res
 
 ![img.png](img.png)
 
-
 ## Requirements
+
 - Python 3.10+
 - `requests` library
 - `web3` (only for Engagement High Signal address checksum)
 
 Install (example):
+
 ```bash
 python -m venv venv && . venv/bin/activate
 pip install -r requirements.txt
 ```
 
 ## Usage
-Script asks for manual input for certain scores (Human Passport, High Signal, Discord/X presence). 
+
+Script asks for manual input for certain scores (Human Passport, High Signal, Discord/X presence).
 
 Run any category or the orchestrator with one or more addresses separated by space.
+
 ```bash
 # Orchestrator (runs all categories)
 python main.py 0xabc... 0xdef...
@@ -38,12 +42,15 @@ python humanity/main.py 0xabc... 0xdef...
 ```
 
 ## How Data Is Collected
+
 - Static, curated snapshots (as-of a date or block):
+
   - Community lists (Experience): [EthStaker](https://github.com/ethstaker/solo-stakers) and [StakeCat](https://github.com/Stake-Cat/Solo-Stakers/tree/main) solo-staker CSVs are compiled from community sources and reflect membership as of the collection date.
   - SSV Verified operators (Experience) are collected on a specific date with exclusion of Pro operators.
   - SDVTM participants (Experience): a snapshot list, not expected to change.
 
 - On-chain, block-stamped reports (as-of a block):
+
   - Aragon voters (Engagement): a compiled CSV of addresses and vote counts.
   - Obol Techne tiers (Experience): Base, Bronze and Silver NFT holders.
   - Circles group members (Humanity): a snapshot list of eligible addresses.
@@ -56,13 +63,15 @@ python humanity/main.py 0xabc... 0xdef...
   - High Signal (Engagement): if `HIGH_SIGNAL_API_KEY` is set, activity scores are queried for a configured date range; otherwise the operator provides the score manually.
   - Gitcoin Passport (Humanity): if `HUMAN_PASSPORT_API_KEY` is set, the passport score is fetched live; otherwise the operator provides the score manually.
 
-
 ## Environment Variables (optional)
+
 - `HIGH_SIGNAL_API_KEY` (Engagement): Enables automatic High Signal lookups; otherwise manual input is prompted.
 - `HUMAN_PASSPORT_API_KEY` (Humanity): Enables automatic Gitcoin Passport score; otherwise manual input is prompted.
 
 ## Tests
+
 Pytest suites are included and fully offline via mocks.
+
 ```bash
 # Install pytest if not already done
 pip install pytest
