+++ ics-assessment/experience/_get_ssv_operators.py
