+++ ics-assessment/experience/eligible_node_operators_hoodi.json
@@ -1,0 +1,49 @@
+[
+  "0",
+  "10",
+  "103",
+  "106",
+  "108",
+  "109",
+  "110",
+  "12",
+  "14",
+  "15",
+  "16",
+  "17",
+  "18",
+  "19",
+  "2",
+  "20",
+  "22",
+  "23",
+  "25",
+  "26",
+  "27",
+  "33",
+  "37",
+  "40",
+  "41",
+  "48",
+  "49",
+  "5",
+  "52",
+  "55",
+  "59",
+  "60",
+  "61",
+  "64",
+  "69",
+  "74",
+  "77",
+  "79",
+  "8",
+  "85",
+  "88",
+  "9",
+  "91",
+  "92",
+  "93",
+  "94",
+  "98"
+]
