+++ ics-assessment/experience/node_operator_owners_hoodi.json
