+++ ics-assessment/experience/node_operator_owners_mainnet.json
