+++ ics-assessment/humanity/_collect_circles.py
