+++ ics-assessment/humanity/main.py
