+++ ics-assessment/tests/test_humanity_main.py
